Use backendUrl store variable directly
