+++ src/modules/prem-audio/api/generateTranscriptions.ts
@@ -1,11 +1,15 @@
 import axios from "axios";
 import { generateUrl } from "shared/helpers/utils";
-import { getBackendUrlFromStore } from "shared/store/setting";
 
+import useSettingStore from "../../../shared/store/setting";
 import type { TranscriptionsGeneration } from "../types";
 
 const generateTranscriptions = async (port: number, data: TranscriptionsGeneration) => {
-  const backendUrl = generateUrl(getBackendUrlFromStore(), port, "v1/audio/transcriptions");
+  const backendUrl = generateUrl(
+    useSettingStore.getState().backendUrl,
+    port,
+    "v1/audio/transcriptions",
+  );
   const formData = new FormData();
   formData.append("file", data.file);
   formData.append("model", data.model);
