+++ src/modules/prem-image/api/generateImage.ts
@@ -1,11 +1,15 @@
 import axios from "axios";
 import { generateUrl } from "shared/helpers/utils";
-import { getBackendUrlFromStore } from "shared/store/setting";
 
+import useSettingStore from "../../../shared/store/setting";
 import type { ImageGeneration } from "../types";
 
 const generateImage = async (port: number, data: ImageGeneration) => {
-  const backendUrl = generateUrl(getBackendUrlFromStore(), port, "v1/images/generations");
+  const backendUrl = generateUrl(
+    useSettingStore.getState().backendUrl,
+    port,
+    "v1/images/generations",
+  );
 
   return axios.post(`${backendUrl}`, data, {
     headers: {
