+++ src/modules/prem-text-audio/api/generateAudio.ts
@@ -1,11 +1,15 @@
 import axios from "axios";
 import { generateUrl } from "shared/helpers/utils";
-import { getBackendUrlFromStore } from "shared/store/setting";
 
+import useSettingStore from "../../../shared/store/setting";
 import type { AudioGenerationData } from "../types";
 
 const generateAudio = async (port: number, data: AudioGenerationData) => {
-  const backendUrl = generateUrl(getBackendUrlFromStore(), port, "v1/audio/generation");
+  const backendUrl = generateUrl(
+    useSettingStore.getState().backendUrl,
+    port,
+    "v1/audio/generation",
+  );
 
   return axios.post(`${backendUrl}`, data, {
     headers: {
