+++ src/modules/prem-upscaler/api/generateUpscalerImage.ts
@@ -1,11 +1,11 @@
 import axios from "axios";
 import { generateUrl } from "shared/helpers/utils";
-import { getBackendUrlFromStore } from "shared/store/setting";
 
+import useSettingStore from "../../../shared/store/setting";
 import type { ImageGeneration } from "../types";
 
 const generateUpscalerImage = async (port: number, data: ImageGeneration) => {
-  const backendUrl = generateUrl(getBackendUrlFromStore(), port, "v1/images/upscale");
+  const backendUrl = generateUrl(useSettingStore.getState().backendUrl, port, "v1/images/upscale");
   const formData = new FormData();
   formData.append("image", data.image);
   formData.append("prompt", data.prompt);
