+++ src/modules/settings/components/BackendUrl.tsx
@@ -5,8 +5,7 @@ import { toast } from "react-toastify";
 import { Tooltip } from "react-tooltip";
 import PrimaryButton from "shared/components/PrimaryButton";
 import ResetIcon from "shared/components/ResetIcon";
-import { getBackendUrl } from "shared/helpers/utils";
-import useSettingStore from "shared/store/setting";
+import useSettingStore, { getBackendUrl } from "shared/store/setting";
 
 import checkHealth from "../api/checkHealth";
