+++ src/shared/store/setting.ts
@@ -1,8 +1,19 @@
 import storage from "shared/helpers/custom-storage";
-import { getBackendUrl } from "shared/helpers/utils";
+import { isBackendSet, isPackaged } from "shared/helpers/utils";
 import type { SettingStore } from "shared/types";
 import { create } from "zustand";
 import { createJSONStorage, devtools, persist } from "zustand/middleware";
+
+export const getBackendUrl = () => {
+  let backendURL = "http://localhost:54321";
+  if (isBackendSet()) {
+    backendURL = import.meta.env.VITE_BACKEND_URL;
+  }
+  if (isPackaged()) {
+    backendURL = `${window.location.protocol}//${window.location.host}/api/`;
+  }
+  return backendURL;
+};
 
 const useSettingStore = create<SettingStore>()(
   devtools(
@@ -51,7 +62,5 @@ const useSettingStore = create<SettingStore>()(
     { name: "store", store: "setting" },
   ),
 );
-
-export const getBackendUrlFromStore = (): string => useSettingStore.getState().backendUrl;
 
 export default useSettingStore;
